+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.1.1 (2023-10-27)
+
+* initial release
+
 ## 0.1.0 (2023-10-26)
 
 * project started